Move `instrument` to debugger Verbose code
